backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-filter==24.3
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.1
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1
rapidfuzz==3.10.1
gunicorn==23.0.0
python-dotenv==1.0.1
Pillow==11.0.0
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-mock==3.14.0
factory-boy==3.3.1
